Cleaned fake data for mortalityregistration

Cleaned fake data for mortalityregistration

Usage

data_fake_nowcasting_nation_aggregated

Format

cut_doe

First date of every week

n_death

Number of true deaths this week

n0_0

Number of registrations within the current week

n0_1

Number of registrations within the current and previous week

p0_1

Percentile of registrations within the current and previous week

n0_2

Number of registrations within the 2 last weeks and the current week

p0_2

Percentile of registrations within the current and prvious 2 weeks

n0_3

Number of registrations within the 3 last weeks and the current week

p0_3

Percentile of registrations within the current and prvious 3 weeks

n0_4

Number of registrations within the 4 weeks and the current week

p0_4

Percentile of registrations within the current and prvious 4 weeks

n0_5

Number of registrations within the 5 weeks and the current week

p0_5

Percentile of registrations within the current and prvious 5 weeks

n0_6

Number of registrations within the 6 weeks and the current week

p0_6

Percentile of registrations within the current and prvious 6 weeks

n0_7

Number of registrations within the 7 weeks and the current week

p0_7

Percentile of registrations within the current and prvious 7 weeks

n0_8

Number of registrations within the 8 weeks and the current week

p0_8

Percentile of registrations within the current and prvious 8 weeks

n0_9

Number of registrations within the 9 weeks and the current week

p0_9

Percentile of registrations within the current and prvious 9 weeks

n0_10

Number of registrations within the 10 weeks and the current week

p0_10

Percentile of registrations within the current and prvious 10 weeks

n0_11

Number of registrations within the 11 weeks and the current week

p0_11

Percentile of registrations within the current and prvious 11 weeks

n0_12

Number of registrations within the 12 weeks and the current week

p0_12

Percentile of registrations within the current and prvious 12 weeks

n0_13

Number of registrations within the 13 weeks and the current week

p0_13

Percentile of registrations within the current and prvious 13 weeks

n0_14

Number of registrations within the 14 weeks and the current week

p0_14

Percentile of registrations within the current and prvious 14 weeks

n0_15

Number of registrations within the 15 weeks and the current week

p0_15

Percentile of registrations within the current and prvious 15 weeks