Baseline estimator.

Baseline estimator for mortality. If a random effect is present the negative binomial distribution is used. Otherwise a quisipoisson is fitted to the data. We use simmulations to generate n_sim responses for each row in the dataset to create a baseline.

baseline_est(
  data_train,
  data_predict,
  n_sim = 1000,
  fixef,
  ranef,
  response,
  offset
)

Arguments

data_train
Data to train on.

data_predict
Data to predict on

n_sim
number of simulations to preform. Default setting is n_sim = 1000

fixef
Fixed effekts

ranef
Random effekts, default is NULL

response
Response

offset
Offset, can be NULL

Value

Residualplots for all ncor_i and some evaluationmetrixs for each of them as well as a plot containing credible intervals using the simulation

Details

For more details see the help vignette: vignette("nowcast", package="attrib")

Examples

if (FALSE) {
data <- as.data.table(data_fake_nowcasting_county_aggregated)
n_sim <- 100
fixef <- "sin(2 * pi * (week) / 53) + cos(2 * pi * (week ) / 53) + year"
ranef <- "(1|location_code)"
response <- "n_death"
data_train <- data[cut_doe< "2019-06-30"]
data_predict <- data
offset <- "log(pop)"
baseline_est(data_train, data_predict, n_sim = 1000, fixef, ranef, response, offset)
}
data <- data.table::as.data.table(data_fake_nowcasting_county_aggregated)
data <- data[location_code == "county03"]
n_sim <- 100
fixef <- "sin(2 * pi * (week) / 53) + cos(2 * pi * (week ) / 53) + year"
ranef <- NULL
response <- "n_death"
data_train <- data[cut_doe< "2019-06-30"]
data_predict <- data
offset <- "log(pop)"
baseline_est(data_train, data_predict, n_sim = 1000, fixef, ranef, response, offset)

#> $simulations
#>            cut_doe location_code n_death n0_0 p0_0 n0_1 p0_1 n0_2 p0_2 n0_3
#>      1: 2018-01-01      county03      10    0    0    4  0.4    9  0.9   10
#>      2: 2018-01-01      county03      10    0    0    4  0.4    9  0.9   10
#>      3: 2018-01-01      county03      10    0    0    4  0.4    9  0.9   10
#>      4: 2018-01-01      county03      10    0    0    4  0.4    9  0.9   10
#>      5: 2018-01-01      county03      10    0    0    4  0.4    9  0.9   10
#>     ---                                                                    
#> 102996: 2019-12-16      county03       1    0    0   NA   NA   NA   NA   NA
#> 102997: 2019-12-16      county03       1    0    0   NA   NA   NA   NA   NA
#> 102998: 2019-12-16      county03       1    0    0   NA   NA   NA   NA   NA
#> 102999: 2019-12-16      county03       1    0    0   NA   NA   NA   NA   NA
#> 103000: 2019-12-16      county03       1    0    0   NA   NA   NA   NA   NA
#>         p0_3 n0_4 p0_4 n0_5 p0_5 week year    pop id_row sim_id sim_value
#>      1:    1   10    1   10    1    1 2018 664234      1      1        10
#>      2:    1   10    1   10    1    1 2018 664234      1      2        12
#>      3:    1   10    1   10    1    1 2018 664234      1      3         8
#>      4:    1   10    1   10    1    1 2018 664234      1      4        14
#>      5:    1   10    1   10    1    1 2018 664234      1      5         9
#>     ---                                                                  
#> 102996:   NA   NA   NA   NA   NA   51 2019 672069    103    996        16
#> 102997:   NA   NA   NA   NA   NA   51 2019 672069    103    997        11
#> 102998:   NA   NA   NA   NA   NA   51 2019 672069    103    998         7
#> 102999:   NA   NA   NA   NA   NA   51 2019 672069    103    999        16
#> 103000:   NA   NA   NA   NA   NA   51 2019 672069    103   1000         9
#>                  type
#>      1: quasi_poisson
#>      2: quasi_poisson
#>      3: quasi_poisson
#>      4: quasi_poisson
#>      5: quasi_poisson
#>     ---              
#> 102996: quasi_poisson
#> 102997: quasi_poisson
#> 102998: quasi_poisson
#> 102999: quasi_poisson
#> 103000: quasi_poisson
#> 
#> $aggregated
#>         cut_doe location_code n_death n0_0 p0_0 n0_1       p0_1 n0_2      p0_2
#>   1: 2018-01-01      county03      10    0    0    4 0.40000000    9 0.9000000
#>   2: 2018-01-08      county03       6    0    0    0 0.00000000    3 0.5000000
#>   3: 2018-01-15      county03      16    0    0    3 0.18750000   11 0.6875000
#>   4: 2018-01-22      county03      12    0    0    2 0.16666667   11 0.9166667
#>   5: 2018-01-29      county03       8    0    0    1 0.12500000    6 0.7500000
#>  ---                                                                          
#>  99: 2019-11-18      county03      16    0    0    3 0.18750000   14 0.8750000
#> 100: 2019-11-25      county03      12    0    0    1 0.08333333   10 0.8333333
#> 101: 2019-12-02      county03      14    0    0    1 0.07142857   11 0.7857143
#> 102: 2019-12-09      county03       9    0    0    3 0.33333333   NA        NA
#> 103: 2019-12-16      county03       1    0    0   NA         NA   NA        NA
#>      n0_3      p0_3 n0_4 p0_4 n0_5 p0_5 week year    pop id_row          type
#>   1:   10 1.0000000   10    1   10    1    1 2018 664234      1 quasi_poisson
#>   2:    6 1.0000000    6    1    6    1    2 2018 664234      2 quasi_poisson
#>   3:   16 1.0000000   16    1   16    1    3 2018 664234      3 quasi_poisson
#>   4:   12 1.0000000   12    1   12    1    4 2018 664234      4 quasi_poisson
#>   5:    8 1.0000000    8    1    8    1    5 2018 664234      5 quasi_poisson
#>  ---                                                                         
#>  99:   16 1.0000000   16    1   NA   NA   47 2019 672069     99 quasi_poisson
#> 100:   11 0.9166667   NA   NA   NA   NA   48 2019 672069    100 quasi_poisson
#> 101:   NA        NA   NA   NA   NA   NA   49 2019 672069    101 quasi_poisson
#> 102:   NA        NA   NA   NA   NA   NA   50 2019 672069    102 quasi_poisson
#> 103:   NA        NA   NA   NA   NA   NA   51 2019 672069    103 quasi_poisson
#>      median.sim_value q025.sim_value q975.sim_value
#>   1:               13          6.000         21.000
#>   2:               13          7.000         21.000
#>   3:               13          6.975         20.025
#>   4:               13          6.000         21.000
#>   5:               13          6.000         21.000
#>  ---                                               
#>  99:               12          6.000         20.025
#> 100:               12          6.000         21.000
#> 101:               13          6.000         20.000
#> 102:               13          6.000         20.000
#> 103:               13          6.000         21.000
#> 
#> $fit
#> 
#> Call:  stats::glm(formula = stats::as.formula(formula), family = "quasipoisson", 
#>     data = data_train)
#> 
#> Coefficients:
#>             (Intercept)  sin(2 * pi * (week)/53)  cos(2 * pi * (week)/53)  
#>               -28.25089                  0.04186                  0.02288  
#>                    year                 log(pop)  
#>                 0.01525                       NA  
#> 
#> Degrees of Freedom: 77 Total (i.e. Null);  74 Residual
#> Null Deviance:	    56.67 
#> Residual Deviance: 55.39 	AIC: NA
#>