Estimates simulations of expected responses

For each exposure the dataset is copied and the original value replaced by the referance value. Then the sim function is used to generate 500 simulations of expected responses for each row. Finaly the dataset is transformed to obtain expected response for original and referance values of the given exposures for each original row of the dataset.

est_attrib(fit, data, exposures)

Arguments

fit
A model fit constructed by fit_attrib

data
The observed data

exposures
The exposures that will get reference expected mortalities

For more details see the help vignette: vignette("intro", package="attrib")

Value

Dataset with expected responses for all simulations including excpected responses given the exposure referance values

Details

For more details see the help vignette: vignette("intro", package="attrib")

Examples

response <- "deaths"
fixef <- "pr100_ili_lag_1 + sin(2 * pi * (week - 1) / 52) + cos(2 * pi * (week - 1) / 52)"
ranef <- " (pr100_ili_lag_1| season)"
offset <- "log(pop)"

data = attrib::data_fake_nation

fit <- fit_attrib(data = data, response = response, fixef = fixef, ranef = ranef, offset = offset)

#> boundary (singular) fit: see ?isSingular

exposures <- c(pr100_ili_lag_1 = 0)
new_data <- est_attrib(fit, data, exposures)

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%

new_data[]

#>         location_code week    season year    yrwk    x     pop pr100_ili
#>      1:        norway    1 2009/2010 2010 2010-01 24.0 5367580 1.9293240
#>      2:        norway    1 2010/2011 2011 2011-01 24.0 5367580 1.9219890
#>      3:        norway    1 2011/2012 2012 2012-01 24.0 5367580 1.3976273
#>      4:        norway    1 2012/2013 2013 2013-01 24.0 5367580 1.7592660
#>      5:        norway    1 2013/2014 2014 2014-01 24.0 5367580 1.1564445
#>     ---                                                                 
#> 286996:        norway   52 2018/2019 2018 2018-52 23.0 5367580 1.4134848
#> 286997:        norway   52 2019/2020 2019 2019-52 23.0 5367580 1.7789371
#> 286998:        norway   52 2020/2021 2020 2020-52 23.0 5367580 0.5308053
#> 286999:        norway   53 2009/2010 2009 2009-53 23.5 5367580 1.9197015
#> 287000:        norway   53 2015/2016 2015 2015-53 23.5 5367580 1.8189595
#>         pr100_ili_lag_1 temperature temperature_high deaths  id sim_id
#>      1:       1.9197015  -3.9131137                0    942   1      1
#>      2:       1.8839310   2.1005689                0    969   2      1
#>      3:       1.1673965  -9.2338013                0    920   3      1
#>      4:       1.5918497  -7.0780783                0    889   4      1
#>      5:       0.9280684  -3.2695932                0    867   5      1
#>     ---                                                               
#> 286996:       1.1806417   0.9522693                0    860 570    500
#> 286997:       1.6096489   7.0470121                0    918 571    500
#> 286998:       0.3778115  -3.6855952                0    915 572    500
#> 286999:       0.0000000   2.8811622                0    857 573    500
#> 287000:       1.7563971   5.0585632                0    943 574    500
#>         exp_mort_observed exp_mort_pr100_ili_lag_1=0
#>      1:          918.8242                   859.2425
#>      2:          950.5683                   855.0132
#>      3:          933.2312                   873.8763
#>      4:          869.4745                   851.1139
#>      5:          882.1646                   850.5451
#>     ---                                             
#> 286996:          843.0222                   843.3506
#> 286997:          942.8308                   875.3058
#> 286998:          863.3456                   852.6389
#> 286999:          881.1744                   881.1744
#> 287000:          902.0270                   839.8431