Data fit

Data fit using glmer from lme4 with family poisson to fit the dataset with the given formula.

fit_attrib(
  data,
  response,
  fixef,
  ranef,
  dist_family = "poisson",
  offset = NULL
)

Arguments

data
The observed data to be fitted.

response
The response

fixef
The fixed effects

ranef
The random effects

dist_family
Family

offset
The offsets.

Value

The model fit of the data with additional attributes offset, response and fit_fix. Offset and response are the same as in the input and fit_fix is the linear model of the fix effects.

For more details see the help vignette: vignette("intro", package="attrib")

Examples

response <- "deaths"

fixef <- "pr100_ili_lag_1 + sin(2 * pi * (week - 1) / 52) + cos(2 * pi * (week - 1) / 52)"
ranef <- " (pr100_ili_lag_1| season)"
offset <- "log(pop)"

data <- attrib::data_fake_nation

fit_attrib(data = data, response = response, fixef = fixef, ranef = ranef, offset = offset)

#> boundary (singular) fit: see ?isSingular

#> Generalized linear mixed model fit by maximum likelihood (Laplace
#>   Approximation) [glmerMod]
#>  Family: poisson  ( log )
#> Formula: deaths ~ pr100_ili_lag_1 + sin(2 * pi * (week - 1)/52) + cos(2 *  
#>     pi * (week - 1)/52) + offset(log(pop)) + (pr100_ili_lag_1 |      season)
#>    Data: data
#>       AIC       BIC    logLik  deviance  df.resid 
#>  5664.885  5695.353 -2825.442  5650.885       567 
#> Random effects:
#>  Groups Name            Std.Dev. Corr
#>  season (Intercept)     0.005714     
#>         pr100_ili_lag_1 0.001022 1.00
#> Number of obs: 574, groups:  season, 11
#> Fixed Effects:
#>                 (Intercept)              pr100_ili_lag_1  
#>                  -8.7962692                    0.0536117  
#> sin(2 * pi * (week - 1)/52)  cos(2 * pi * (week - 1)/52)  
#>                  -0.0003147                    0.0425239  
#> optimizer (Nelder_Mead) convergence code: 0 (OK) ; 0 optimizer warnings; 1 lme4 warnings