Generates simulations of expected mortality by simulating the model coefficients.

With the given fit from fit_attrib the function sim, from package arm, is used to generate 500 simulations of all the coefficients, from there respective posterior distributions. This is then used to compute the expected response for all simulations and rows in the input dataset.

sim(fit, data, n_sim, PI = FALSE)

Arguments

fit
A model fit created by fit_attrib

data
The data with either observed values or reference values.

n_sim
Number of simulations

PI
Returns a draw from the quasipoisson distribution with the calculated mean as mean. If false the exponential of the mean is returned.

Value

A dataset with 500 simulations of the expected response for each row in the original dataset.

Details

vignette("intro", package="attrib")

Examples

response <- "deaths"
fixef <- "pr100_ili_lag_1 + sin(2 * pi * (week - 1) / 52) + cos(2 * pi * (week - 1) / 52)"
ranef <- " (pr100_ili_lag_1| season)"
offset <- "log(pop)"

data <- attrib::data_fake_nation

fit <- fit_attrib(data = data, response = response, fixef = fixef, ranef = ranef, offset = offset)

#> boundary (singular) fit: see ?isSingular

n_sim <- 5
sim(fit, data, n_sim)

#>       week    season year    yrwk    x location_code     pop pr100_ili
#>    1:    1 2009/2010 2010 2010-01 24.0         norge 5367580  1.160403
#>    2:    1 2010/2011 2011 2011-01 24.0         norge 5367580  1.164511
#>    3:    1 2011/2012 2012 2012-01 24.0         norge 5367580  1.472767
#>    4:    1 2012/2013 2013 2013-01 24.0         norge 5367580  1.435822
#>    5:    1 2013/2014 2014 2014-01 24.0         norge 5367580  1.402403
#>   ---                                                                 
#> 2866:   52 2017/2018 2017 2017-52 23.0         norge 5367580  1.270134
#> 2867:   52 2018/2019 2018 2018-52 23.0         norge 5367580  1.028794
#> 2868:   52 2019/2020 2019 2019-52 23.0         norge 5367580  1.236302
#> 2869:   53 2009/2010 2009 2009-53 23.5         norge 5367580  1.100072
#> 2870:   53 2015/2016 2015 2015-53 23.5         norge 5367580  1.269054
#>       pr100_ili_lag_1 temperature_high deaths id_row sim_id sim_value
#>    1:       1.1000719                0    891      1      1  899.6451
#>    2:       0.9623231                0    874      2      1  887.7409
#>    3:       1.3356371                0    854      3      1  912.2060
#>    4:       1.2320998                0    885      4      1  907.7964
#>    5:       1.2514772                0    904      5      1  901.3637
#>   ---                                                                
#> 2866:       1.0938242                0    854    570      5  908.0504
#> 2867:       0.8534488                0    893    571      5  885.1736
#> 2868:       1.0609100                0    829    572      5  899.4894
#> 2869:       1.0387242                0    880    573      5  895.6281
#> 2870:       1.1645116                0    912    574      5  896.8174